refactor!: flatten the module structure
traits and clients are now on the top level. Error and response structs are moved to their dedicated modules.
Also, a bunch of doc comments are added so the docs.rs/frankenstein entry has somewhat better docs

BREAKING CHANGE: old modules for them are gone and import is only possible from the top level.
Responses moved from frankenstein::api_traits::MethodResponse to frankenstein::response::MethodResponse (and are also reexported on top level).
