+++ src/objects.rs
@@ -1,9 +1,10 @@
+//! Objects returned or used with the Telegram API.
+
 #![allow(deprecated)]
-use super::api_params::FileUpload;
+use crate::api_params::FileUpload;
+use crate::ParseMode;
 use bon::Builder;
 use serde::{Deserialize, Serialize};
-
-use crate::ParseMode;
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
 #[serde(rename_all = "snake_case")]
