+++ src/parse_mode.rs
@@ -5,6 +5,9 @@ use std::str::FromStr;
 
 use serde::{Deserialize, Serialize};
 
+/// Text Formatting Options
+///
+/// See <https://core.telegram.org/bots/api#formatting-options>
 #[derive(Clone, Copy, Debug, Hash, PartialEq, Eq, Serialize, Deserialize)]
 pub enum ParseMode {
     #[serde(rename = "HTML")]
@@ -18,6 +21,7 @@ pub enum ParseMode {
 
 impl FromStr for ParseMode {
     type Err = ();
+
     fn from_str(s: &str) -> Result<Self, Self::Err> {
         match s {
             "HTML" | "Html" | "html" => Ok(Self::Html),
