+++ src/trait_async.rs
@@ -1,5 +1,3 @@
-use super::EditMessageResponse;
-use super::MethodResponse;
 use crate::api_params::AddStickerToSetParams;
 use crate::api_params::AnswerCallbackQueryParams;
 use crate::api_params::AnswerInlineQueryParams;
@@ -120,6 +118,7 @@ use crate::api_params::UnbanChatSenderChatParams;
 use crate::api_params::UnhideGeneralForumTopicParams;
 use crate::api_params::UnpinAllChatMessagesParams;
 use crate::api_params::UnpinAllForumTopicMessagesParams;
+use crate::api_params::UnpinAllGeneralForumTopicMessagesParams;
 use crate::api_params::UnpinChatMessageParams;
 use crate::api_params::UploadStickerFileParams;
 use crate::objects::BotCommand;
@@ -141,14 +140,14 @@ use crate::objects::MessageId;
 use crate::objects::Poll;
 use crate::objects::SentWebAppMessage;
 use crate::objects::StarTransactions;
+use crate::objects::Sticker;
 use crate::objects::StickerSet;
 use crate::objects::Update;
 use crate::objects::User;
 use crate::objects::UserChatBoosts;
 use crate::objects::UserProfilePhotos;
 use crate::objects::WebhookInfo;
-use crate::Sticker;
-use crate::UnpinAllGeneralForumTopicMessagesParams;
+use crate::response::{EditMessageResponse, MethodResponse};
 use async_trait::async_trait;
 use std::path::PathBuf;
