+++ src/trait_sync.rs
@@ -1,5 +1,3 @@
-use super::EditMessageResponse;
-use super::MethodResponse;
 use crate::api_params::AddStickerToSetParams;
 use crate::api_params::AnswerCallbackQueryParams;
 use crate::api_params::AnswerInlineQueryParams;
@@ -47,6 +45,7 @@ use crate::api_params::GetChatMemberCountParams;
 use crate::api_params::GetChatMemberParams;
 use crate::api_params::GetChatMenuButtonParams;
 use crate::api_params::GetChatParams;
+use crate::api_params::GetCustomEmojiStickersParams;
 use crate::api_params::GetFileParams;
 use crate::api_params::GetGameHighScoresParams;
 use crate::api_params::GetMyCommandsParams;
@@ -119,6 +118,7 @@ use crate::api_params::UnbanChatSenderChatParams;
 use crate::api_params::UnhideGeneralForumTopicParams;
 use crate::api_params::UnpinAllChatMessagesParams;
 use crate::api_params::UnpinAllForumTopicMessagesParams;
+use crate::api_params::UnpinAllGeneralForumTopicMessagesParams;
 use crate::api_params::UnpinChatMessageParams;
 use crate::api_params::UploadStickerFileParams;
 use crate::objects::BotCommand;
@@ -140,15 +140,14 @@ use crate::objects::MessageId;
 use crate::objects::Poll;
 use crate::objects::SentWebAppMessage;
 use crate::objects::StarTransactions;
+use crate::objects::Sticker;
 use crate::objects::StickerSet;
 use crate::objects::Update;
 use crate::objects::User;
 use crate::objects::UserChatBoosts;
 use crate::objects::UserProfilePhotos;
 use crate::objects::WebhookInfo;
-use crate::GetCustomEmojiStickersParams;
-use crate::Sticker;
-use crate::UnpinAllGeneralForumTopicMessagesParams;
+use crate::response::{EditMessageResponse, MethodResponse};
 use std::path::PathBuf;
 
 pub trait TelegramApi {
